Add a couple of real Shurikens
I've also found more than one valid solution, so I also had to create a
mechanism for accepting various different solutions as valid answers.
Not only that, but it also turns out this can take a while to calculate,
so I had to make sure it was behind a parameter wall.