Treat both sides of the shuriken as interchangeable
This leads to a massive performance improvement, as we now only have to
care about half of the shurikens/nodes we needed to. The run times have
not decrease in half (because we weren't hitting all the nodes we now
ignore), but it's still very significant.
A side effect was that the model unit grew too big, so I moved most of the code
out of the header file into a new compilation unit.