Rewrite ThreadSafeQueue::push_nodes for efficiency
The old version was instantiating loads of `Node *` while holding the
lock, so I changed it to instantiate them beforehand. Another sideeffect
is that it then became possible to build the whole chain before grabbing
the lock as well (i.e. all the "x->next = y"), so that's a bonus!