Code Cleanup (assertions)
Removed the unused <assert> include, surrounded the no-op version with
() and changed ensure_m's static_cast to a dynamic_cast to avoid a
downcast warning. Technically the static_cast works perfectly, but since
it's just for the debug builds, the small performance cost should give
us less of a headache than the IDE's warning.

Oh, and also removed an extraneous ; that shouldn't be there