[CMake] Disable RTTI
Also changed the ensure_* functions to use a C-style cast. We could also
have used a static_cast, but then CLion would be littered with warnings
about using a static_cast to downcast, because technically we're casting
a std::basic_ostream<char> to a std::stringstream. A C-style cast
accomplishes the same thing, without triggering that warning.