[ACTIONS] Fix the Actions Ubuntu build
Apparently, SOMETHING changed on the images... or maybe on ubuntu's
packages. One way or the other, the old script just doesn't work
anymore, ubuntu clang builds are now failing with an ld error:

/usr/bin/ld: cannot find -lunwind

I'm not sure what broke, nor if this fix is the "correct" one, but I'm
calling this good enough.